教师学生的用户登录登出保持接口完成

+++ src/main/java/tbs/newgenteacherselect/config/impl/AccessManager.java
@@ -2,6 +2,7 @@ package tbs.newgenteacherselect.config.impl;
 
 import org.springframework.stereotype.Component;
 import tbs.dao.RoleDao;
+import tbs.newgenteacherselect.NetErrorEnum;
 import tbs.utils.AOP.authorize.interfaces.IAccess;
 import tbs.utils.AOP.authorize.model.BaseRoleModel;
 import tbs.utils.redis.IRedisService;
@@ -25,7 +26,7 @@ public class AccessManager implements IAccess {
     long login_timeout = 30;
     TimeUnit login_timeout_unit = TimeUnit.MINUTES;
 
-    public static final String PREFIX = "ACCESS_TOKEN_";
+    public static final String PREFIX = "ACCESS_TOKEN_", SINGLE_LOGIN_PREFIX = "SINGLE_LOGIN_";
 
     @Override
     public BaseRoleModel readRole(String tokenStr) {
@@ -47,12 +48,30 @@ public class AccessManager implements IAccess {
     }
 
     @Override
-    public void put(String token, BaseRoleModel detail) {
-        redisService.set(PREFIX + token, detail, login_timeout, login_timeout_unit);
+    public void put(String token, BaseRoleModel detail) throws Exception {
+
+        if (detail != null&&redisService.hasKey(SINGLE_LOGIN_PREFIX + detail.getUserId())) {
+            throw NetErrorEnum.makeError(NetErrorEnum.Repeated_Login);
+        }
+
+        if (detail == null)
+            detail = readRole(token);
+        if (detail != null) {
+            redisService.set(SINGLE_LOGIN_PREFIX + detail.getUserId(), detail,login_timeout, login_timeout_unit);
+            redisService.set(PREFIX + token, detail, login_timeout, login_timeout_unit);
+            return;
+        }
+        throw NetErrorEnum.makeError(NetErrorEnum.BAD_ROLE);
+
+
     }
 
     @Override
     public void deleteToken(String token) {
+        BaseRoleModel roleModel = this.readRole(token);
+        if (roleModel == null)
+            return;
+        redisService.delete(SINGLE_LOGIN_PREFIX + roleModel.getUserId());
         redisService.delete(PREFIX + token);
     }
 }
